move reordering of stats from SummarizerSingleTensor to Summarizer
