+++ captum/attr/_utils/stat.py
@@ -208,90 +208,3 @@ class Max(GeneralAccumFn):
 class Sum(GeneralAccumFn):
     def __init__(self, name=None):
         super().__init__(name=name, fn=torch.add)
-
-
-class SummarizerSingleTensor:
-    """
-    A simple class that summarizes a single tensor. The basic functionality
-    of this class is two operations .update and .summary
-    """
-
-    class StatHolder:
-        def __init__(self, stats):
-            # We want to want to store two things:
-            # 1. A mapping from a Stat to Stat object (self._stat_to_stat):
-            #    This is to retrieve an existing Stat object for dependency
-            #    resolution, e.g.  Mean needs the Count stat - we want to
-            #    retrieve it in O(1)
-            #
-            # 2. All of the necessary stats, in the correct order,
-            #    to perform an update for each Stat (self.stats) trivially
-
-            # As a reference, the dependency graph for our stats is as follows:
-            # StdDev(x) -> Var(x) -> MSE -> Mean -> Count, for all valid x
-            #
-            # Step 1:
-            #    Ensure we have all the necessary stats
-            #    i.e. ensure we have the dependencies
-            # Step 2:
-            #    Figure out the order to update them
-            dep_order = [StdDev, Var, MSE, Mean, Count]
-
-            stats = set(stats)  # remove dupe stats
-
-            from collections import defaultdict
-
-            stats_by_module = defaultdict(list)
-            for stat in stats:
-                stats_by_module[stat.__class__].append(stat)
-
-            # StdDev is an odd case since it is parameterized, thus
-            # for each StdDev(order) we must ensure there is an associated Var(order)
-            for std_dev in stats_by_module[StdDev]:
-                stat_to_add = Var(order=std_dev.order)
-                stats.add(stat_to_add)
-                stats_by_module[stat_to_add.__class__].append(stat_to_add)
-
-            # For the other modules (deps[1:n-1]): if i exists =>
-            # we want to ensure i...n-1 exists
-            for i, dep in enumerate(dep_order[1:]):
-                if dep in stats_by_module:
-                    stats.update([mod() for mod in dep_order[i + 1 :]])
-                    break
-
-            # Step 2: get the correct order
-            # NOTE: we are sorting via a given topological order
-            sort_order = {mod: i for i, mod in enumerate(dep_order)}
-            sort_order[Min] = -1
-            sort_order[Max] = -1
-            sort_order[Sum] = -1
-
-            stats = list(stats)
-            stats.sort(key=lambda x: sort_order[x.__class__], reverse=True)
-
-            self.stats = stats
-            self.stat_to_stat = {stat: stat for stat in self.stats}
-
-            for stat in stats:
-                stat._other_stats = self
-                stat.init()
-
-        def get(self, stat):
-            if stat not in self.stat_to_stat:
-                return None
-
-            return self.stat_to_stat[stat]
-
-    def __init__(self, stats=None):
-        self._all_stats = SummarizerSingleTensor.StatHolder(stats)
-
-        # this is what we actually want to output
-        self._summary_stats = stats
-
-    def update(self, x=None):
-        for stat in self._all_stats.stats:
-            stat.update(x)
-
-    @property
-    def summary(self):
-        return {stat.name: stat.get() for stat in self._summary_stats}
